Return iterator from get_changes_iter
This brings the method more in line with what is expected in Rust.

Directly returning a struct implementing `Iterator` did not work due to
lifetimes.

`get_changes_iter` should technically only be supported by a `Session`
in a `module_change_subscribe` callback. This will be addressed in a
later commit.